requirements.txt
Django==3.2.5
django-extensions==3.1.3
django-crispy-forms==1.12.0
django-phonenumber-field==5.2.0
ipdb==0.13.9
psycopg2-binary==2.9.1
phonenumbers==8.12.28
